backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
pyjwt
python-multipart
